Refactor log analysis sns code to prepare for s3sns opstool
