+++ internal/compliance/aws_event_processor/processor/handler.go
@@ -38,8 +38,8 @@ import (
 
 	api "github.com/panther-labs/panther/api/lambda/resources/models"
 	"github.com/panther-labs/panther/internal/compliance/snapshot_poller/models/poller"
-	"github.com/panther-labs/panther/internal/log_analysis/datacatalog_updater/process"
 	"github.com/panther-labs/panther/internal/log_analysis/log_processor/sources"
+	"github.com/panther-labs/panther/internal/log_analysis/notify"
 	"github.com/panther-labs/panther/pkg/awsbatch/sqsbatch"
 	"github.com/panther-labs/panther/pkg/oplog"
 )
@@ -132,7 +132,7 @@ func Handle(lc *lambdacontext.LambdaContext, batch *events.SQSEvent) (err error)
 }
 
 func handleLogProcessorCloudTrail(messageBody string, changes map[string]*resourceChange) (ok bool, err error) {
-	notification := &process.S3Notification{}
+	notification := &notify.S3Notification{}
 	if err := jsoniter.UnmarshalFromString(messageBody, notification); err != nil {
 		return false, errors.Wrap(err, "failed to unmarshal record")
 	}
