+++ internal/log_analysis/datacatalog_updater/process/process_test.go
@@ -33,6 +33,7 @@ import (
 
 	"github.com/panther-labs/panther/internal/log_analysis/awsglue"
 	"github.com/panther-labs/panther/internal/log_analysis/log_processor/registry"
+	"github.com/panther-labs/panther/internal/log_analysis/notify"
 	"github.com/panther-labs/panther/pkg/testutils"
 )
 
@@ -154,7 +155,7 @@ func generateLogTablesMock(logTypes ...string) (tables []*glue.TableData) {
 func getEvent(t *testing.T, s3Keys ...string) events.SQSEvent {
 	result := events.SQSEvent{Records: []events.SQSMessage{}}
 	for _, s3Key := range s3Keys {
-		s3Notification := NewS3ObjectPutNotification("bucket", s3Key, 0)
+		s3Notification := notify.NewS3ObjectPutNotification("bucket", s3Key, 0)
 		serialized, err := jsoniter.MarshalToString(s3Notification)
 		require.NoError(t, err)
 		event := events.SQSMessage{
